renamed functions and refactored them as well
